Moved some functions to helpers.py
