+++ bot/cogs/ping_command.py
@@ -1,10 +1,8 @@
-import csv
-import datetime
 from discord import app_commands
 from discord.ext import commands
 import discord
 from ..modals.feedback_modal import FeedbackModal
-from ..claim import Claim
+from bot.helpers import find_case
 
 # Use TYPE_CHECKING to avoid circular import from bot
 from typing import TYPE_CHECKING
@@ -38,9 +36,8 @@ class PingCommand(commands.Cog):
         """
         # Check if user is a lead
         if self.bot.check_if_lead(interaction.user):
-            case = Claim(case_num=case_num, tech_id=user.id)
-            case_row = self.find_case(case_num, user.id)
-            if len(case_row) == 0:
+            case = find_case(case_num, user_id=user.id, pinged=False)
+            if case is None:
                 await interaction.response.send_message(content="Error! Case couldn't be found. Please ensure it's already been checked and reported in the log file.", ephemeral=True)
                 return
 
@@ -54,52 +51,3 @@ class PingCommand(commands.Cog):
                 color=discord.Color.red()
             )
             await interaction.response.send_message(embed=bad_user_embed, ephemeral=True)
-    
-
-    def find_case(self, case_num: str, user_id: int) -> list[str]:
-        """Finds the first case in the log file that matches
-        the provided information and hasn't been pinged.
-
-        Args:
-            case_num (str): The case number of the case that needs to be pinged.
-            user_id (int): The user for the case that needs to be pinged.
-
-        Returns:
-            list[str]: The row from the log file for this case.
-        """
-        case_row = []
-        with open('log.csv', 'r') as f:
-            reader = csv.reader(f)
-            for row in reader:
-                # Save row once found
-                if row[2] == case_num and row[5] != "Pinged" and int(row[3]) == user_id:
-                    case_row = row
-                    break
-        
-        return case_row
-    
-
-    def remove_case(self, user_id: int, case_num: str) -> None:
-        """Removes the first case in the log file that
-        matches the provided information and hasn't been pinged.
-
-        Args:
-            user_id (int): The user for the case that needs to be pinged.
-            case_num (str): The case number of the case that needs to be pinged.
-        """
-        # Collect rows with this case
-        lines = []
-        found_row = False
-        with open('log.csv', 'r') as f:
-            reader = csv.reader(f)
-            for row in reader:
-                # Exclude row once found
-                if not found_row and row[2] == case_num and row[5] != "Pinged" and int(row[3]) == user_id:
-                    found_row = True
-                    continue
-                
-                lines.append(row)
-                    
-        with open('log.csv', 'w', newline='', encoding='utf-8') as f:
-            writer = csv.writer(f)
-            writer.writerows(lines)
