+++ bot/modals/feedback_modal.py
@@ -1,8 +1,8 @@
-import csv
 import discord
 import discord.ui as ui
 from datetime import datetime
 
+from bot.helpers import remove_case
 from bot.views.ping_view import PingView
 from ..claim import Claim
 
@@ -74,7 +74,7 @@ class FeedbackModal(ui.Modal, title='Feedback Form'):
         await interaction.response.send_message(content="Pinged", delete_after=0) # Acknowledge interaction, immediately delete message
 
         # Remove unpinged case from log
-        self.remove_case(self.case.tech_id, self.case.case_num)
+        remove_case(self.case.tech_id, self.case.case_num)
 
         # Update case, re-log it
         old_message_id = self.case.message_id
@@ -89,29 +89,3 @@ class FeedbackModal(ui.Modal, title='Feedback Form'):
 
         # Remove from active cases
         self.bot.remove_case(old_message_id)
-    
-
-    def remove_case(self, user_id: int, case_num: str) -> None:
-        """Removes the first case in the log file that
-        matches the provided information and hasn't been pinged.
-
-        Args:
-            user_id (int): The user for the case that needs to be pinged.
-            case_num (str): The case number of the case that needs to be pinged.
-        """
-        # Collect rows with this case
-        lines = []
-        found_row = False
-        with open('log.csv', 'r') as f:
-            reader = csv.reader(f)
-            for row in reader:
-                # Exclude row once found
-                if not found_row and row[2] == case_num and row[5] != "Pinged" and int(row[3]) == user_id:
-                    found_row = True
-                    continue
-                
-                lines.append(row)
-                    
-        with open('log.csv', 'w', newline='', encoding='utf-8') as f:
-            writer = csv.writer(f)
-            writer.writerows(lines)
